\name{dataDesc-package}


\alias{dataDesc-package}
\alias{dataDesc}

\docType{package}

\title{
  dataDesc
}

\description{
  Data descriptions tools
}

\details{
  \tabular{ll}{
    Package: \tab  dataDesc \cr
    Type: \tab
    Package\cr
    Version: \tab  1.0.0 \cr
    Date: \tab  01/10/2022 \cr
    License: \tab neant \cr
    LazyLoad: \tab yes\cr
  }


}

\author{
  Hugo Marthinet  <h.marthinet@gmail.com> \cr
}

\references{

}

\seealso{
  Other packages by the same author :

  \code{\link{epiplus}}, \code{\link{lazyMe}}, \code{\link{manageR}},
  \code{\link{surveyscores}}, \code{\link{sampleSize}}
}

\examples{

}
