\name{lazyMe-package}

\alias{lazyMe-package}
\alias{lazyMe}

\docType{package}

\title{
  lazyMe
}

\description{
  Personnal coding shorcuts of a lazy data scientist.
}

\details{
  \tabular{ll}{
    Package: \tab  lazyMe \cr
    Type: \tab
    Package\cr
    Version: \tab  0.2.0 \cr
    Date: \tab  01/10/2022 \cr
    License: \tab neant \cr
    LazyLoad: \tab yes\cr
  }


}

\author{
  Hugo Marthinet  <h.marthinet@gmail.com> \cr
}

\references{

}

\seealso{
  Other packages from the same author :

  \code{\link{epiplus}}, \code{\link{dataDesc}}, \code{\link{manageR}},
  \code{\link{surveyscores}}, \code{\link{sampleSize}}
}

\examples{

}
