\name{lazyMe-package}

\alias{lazyMe-package}
\alias{lazyMe}

\docType{package}

\title{
  lazyMe
}

\description{
  Ensemble de traitements fonctions regulierement utilisees dans notre usage de R
  (code, transformation de donnees, acces aux variables R...),
  mises sous formes de package afin de garantir un acces pratique.
}

\details{
  \tabular{ll}{
    Package: \tab  lazyMe \cr
    Type: \tab
    Package\cr
    Version: \tab  0.1.0 \cr
    Date: \tab  01/10/2022 \cr
    License: \tab neant \cr
    LazyLoad: \tab yes\cr
  }


}

\author{
  Hugo Marthinet  <h.marthinet@gmail.com> \cr
}

\references{

}

\seealso{
  Autres packages developpes au CESPA :

  \code{\link{epiplus}}, \code{\link{MDBprod}}, \code{\link{manageR}},
  \code{\link{surveyscores}}, \code{\link{NSN}}
}

\examples{

}
